Borrowing and returning books are working fully

+++ src/main/java/com/library/repository/ItemRepository.java
@@ -3,5 +3,5 @@ package com.library.repository;
 import com.library.domain.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemRepository extends CrudRepository<Item, Long> {
+public interface ItemRepository extends CrudRepository<Item, Integer> {
 }
